removed CellCellMesh, passed to a cellcell module
